+++ packages/search/src/SearchPopover.tsx
@@ -52,19 +52,20 @@ export const SearchPopover: React.FC<{
         setKeyword,
     } = useSearch(store);
 
+    const performSearch = (cb?: () => void) => {
+        setIsQuerying(true);
+        // eslint-disable-next-line @typescript-eslint/no-unused-vars
+        search().then((_) => {
+            setIsQuerying(false);
+            setSearchDone(true);
+            cb && cb();
+        });
+    };
+
     const onKeydownSearch = (e: React.KeyboardEvent<HTMLInputElement>): void => {
         // Press the Enter key
         if (e.key === 'Enter' && keyword) {
-            if (searchDone) {
-                jumpToNextMatch();
-            } else {
-                setIsQuerying(true);
-                // eslint-disable-next-line @typescript-eslint/no-unused-vars
-                search().then((_) => {
-                    setIsQuerying(false);
-                    setSearchDone(true);
-                });
-            }
+            searchDone ? jumpToNextMatch() : performSearch();
         }
     };
 
@@ -88,11 +89,21 @@ export const SearchPopover: React.FC<{
         setKeyword(value);
     };
 
+    React.useEffect(() => {
+        const initialKeyword = store.get('initialKeyword');
+        if (initialKeyword && initialKeyword.length === 1 && keyword) {
+            performSearch(() => {
+                store.update('initialKeyword', []);
+            });
+        }
+    }, []);
+
     const searchLabel =
         l10n && l10n.search ? ((l10n.search as LocalizationMap).enterToSearch as string) : 'Enter to search';
     const previousMatchLabel =
         l10n && l10n.search ? ((l10n.search as LocalizationMap).previousMatch as string) : 'Previous match';
     const nextMatchLabel = l10n && l10n.search ? ((l10n.search as LocalizationMap).nextMatch as string) : 'Next match';
+    const closeButtonLabel = l10n && l10n.search ? ((l10n.search as LocalizationMap).close as string) : 'Close';
 
     return (
         <div className="rpv-search__popover">
@@ -124,6 +135,7 @@ export const SearchPopover: React.FC<{
             <label className="rpv-search__popover-label">
                 <input
                     className="rpv-search__popover-label-checkbox"
+                    data-testid="search__popover-match-case"
                     checked={matchCase}
                     type="checkbox"
                     onChange={onChangeMatchCase}
@@ -134,6 +146,7 @@ export const SearchPopover: React.FC<{
                 <input
                     className="rpv-search__popover-label-checkbox"
                     checked={wholeWords}
+                    data-testid="search__popover-whole-words"
                     type="checkbox"
                     onChange={onChangeWholeWords}
                 />{' '}
@@ -181,9 +194,7 @@ export const SearchPopover: React.FC<{
                         'rpv-search__popover-footer-button--rtl': isRtl,
                     })}
                 >
-                    <Button onClick={onClose}>
-                        {l10n && l10n.search ? ((l10n.search as LocalizationMap).close as string) : 'Close'}
-                    </Button>
+                    <Button onClick={onClose}>{closeButtonLabel}</Button>
                 </div>
             </div>
         </div>